refactor: move commit syncers into a single module
